+++ src/endpoints/invoices.js
@@ -1,0 +1,13 @@
+module.exports = (function () {
+  var Resource = require('./resource.js')
+  function Invoices (master) {
+    Resource.call(this, master)
+
+    this.name = 'invoices'
+    this.endpoint = '/' + this.name
+  }
+
+  Invoices.prototype = new Resource()
+
+  return Invoices
+})()