+++ src/endpoints/taxes.js
@@ -1,0 +1,13 @@
+module.exports = (function () {
+  var Resource = require('./resource.js')
+  function Taxes (master) {
+    Resource.call(this, master)
+
+    this.name = 'taxes'
+    this.endpoint = '/' + this.name
+  }
+
+  Taxes.prototype = new Resource()
+
+  return Taxes
+})()